automations/requirements.txt
functions-framework==3.*
google-api-python-client==2.*
google-auth==2.*
reportlab==4.*
PyYAML==6.*
